Design : 권한 설정을 위한 Acitivity 따로 구현
