+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">Kotlin_Photo_Album</string>
     <string name="open_album">OPEN ALBUM</string>
+    <string name="permission_btn">권한 설정으로</string>
 </resources>
